+++ gms/api/src/main/snapshot/com.linkedin.chart.charts.snapshot.json
@@ -361,32 +361,33 @@
     } ]
   }, "com.linkedin.common.TagAssociation", "com.linkedin.common.TagUrn", "com.linkedin.common.Time", "com.linkedin.common.Url", "com.linkedin.common.Urn", {
     "type" : "record",
+    "name" : "EntityRelationship",
+    "namespace" : "com.linkedin.common.relationships",
+    "doc" : "Downstream lineage information about a dataset including the source reporting the lineage",
+    "fields" : [ {
+      "name" : "created",
+      "type" : "com.linkedin.common.AuditStamp",
+      "doc" : "Audit stamp containing who reported the lineage and when",
+      "optional" : true
+    }, {
+      "name" : "entity",
+      "type" : "com.linkedin.common.Urn",
+      "doc" : "The downstream dataset the lineage points to"
+    } ]
+  }, {
+    "type" : "record",
     "name" : "GenericLineage",
-    "namespace" : "com.linkedin.common.lineage",
+    "namespace" : "com.linkedin.common.relationships",
     "doc" : "Downstream lineage of a dataset",
     "fields" : [ {
       "name" : "entities",
       "type" : {
         "type" : "array",
-        "items" : {
-          "type" : "record",
-          "name" : "GenericLineageRelationship",
-          "doc" : "Downstream lineage information about a dataset including the source reporting the lineage",
-          "fields" : [ {
-            "name" : "auditStamp",
-            "type" : "com.linkedin.common.AuditStamp",
-            "doc" : "Audit stamp containing who reported the lineage and when",
-            "optional" : true
-          }, {
-            "name" : "entity",
-            "type" : "com.linkedin.common.Urn",
-            "doc" : "The downstream dataset the lineage points to"
-          } ]
-        }
+        "items" : "EntityRelationship"
       },
       "doc" : "List of related entities"
     } ]
-  }, "com.linkedin.common.lineage.GenericLineageRelationship", {
+  }, {
     "type" : "record",
     "name" : "Chart",
     "namespace" : "com.linkedin.dashboard",
@@ -882,7 +883,7 @@
           "name" : "genericDownstreamLineage",
           "namespace" : "com.linkedin.chart",
           "path" : "/charts/{key}/genericDownstreamLineage",
-          "schema" : "com.linkedin.common.lineage.GenericLineage",
+          "schema" : "com.linkedin.common.relationships.GenericLineage",
           "doc" : "Rest.li entry point: /chart/{chartKey}/genericDownstreamLineage\n\ngenerated from: com.linkedin.metadata.resources.dashboard.GenericChartDownstreamLineageResource",
           "simple" : {
             "supports" : [ "get" ],
