+++ gms/api/src/main/snapshot/com.linkedin.dataset.datasets.snapshot.json
@@ -386,32 +386,33 @@
     } ]
   }, "com.linkedin.common.fieldtransformer.TransformationType", "com.linkedin.common.fieldtransformer.UDFTransformer", {
     "type" : "record",
+    "name" : "EntityRelationship",
+    "namespace" : "com.linkedin.common.relationships",
+    "doc" : "Downstream lineage information about a dataset including the source reporting the lineage",
+    "fields" : [ {
+      "name" : "created",
+      "type" : "com.linkedin.common.AuditStamp",
+      "doc" : "Audit stamp containing who reported the lineage and when",
+      "optional" : true
+    }, {
+      "name" : "entity",
+      "type" : "com.linkedin.common.Urn",
+      "doc" : "The downstream dataset the lineage points to"
+    } ]
+  }, {
+    "type" : "record",
     "name" : "GenericLineage",
-    "namespace" : "com.linkedin.common.lineage",
+    "namespace" : "com.linkedin.common.relationships",
     "doc" : "Downstream lineage of a dataset",
     "fields" : [ {
       "name" : "entities",
       "type" : {
         "type" : "array",
-        "items" : {
-          "type" : "record",
-          "name" : "GenericLineageRelationship",
-          "doc" : "Downstream lineage information about a dataset including the source reporting the lineage",
-          "fields" : [ {
-            "name" : "auditStamp",
-            "type" : "com.linkedin.common.AuditStamp",
-            "doc" : "Audit stamp containing who reported the lineage and when",
-            "optional" : true
-          }, {
-            "name" : "entity",
-            "type" : "com.linkedin.common.Urn",
-            "doc" : "The downstream dataset the lineage points to"
-          } ]
-        }
+        "items" : "EntityRelationship"
       },
       "doc" : "List of related entities"
     } ]
-  }, "com.linkedin.common.lineage.GenericLineageRelationship", {
+  }, {
     "type" : "record",
     "name" : "Dataset",
     "namespace" : "com.linkedin.dataset",
@@ -1625,7 +1626,7 @@
           "name" : "genericDownstreamLineage",
           "namespace" : "com.linkedin.dataset",
           "path" : "/datasets/{dataset}/genericDownstreamLineage",
-          "schema" : "com.linkedin.common.lineage.GenericLineage",
+          "schema" : "com.linkedin.common.relationships.GenericLineage",
           "doc" : "Rest.li entry point: /datasets/{datasetKey}/genericDownstreamLineage\n\ngenerated from: com.linkedin.metadata.resources.dataset.GenericDownstreamLineageResource",
           "simple" : {
             "supports" : [ "get" ],
